+++ tests/testOpTrainClassifierBlocked.py
@@ -8,7 +8,7 @@ from lazyflow.classifiers import VigraRfLazyflowClassifierFactory, VigraRfLazyfl
 class TestOpTrainRandomForestBlocked(object):
     
     def testBasic(self):
-        features = numpy.indices( (100,100) ).astype(numpy.float) + 0.5
+        features = numpy.indices( (100,100) ).astype(numpy.float32) + 0.5
         features = numpy.rollaxis(features, 0, 3)
         features = vigra.taggedView(features, 'xyc')
         labels = numpy.zeros( (100,100,1), dtype=numpy.uint8 )
